Renamed errors.py -> exceptions.py
